Function fail::cfg_callback

pub fn cfg_callback<S, F>(name: S, f: F) -> Result<(), String> where
    S: Into<String>,
    F: Fn() + Send + Sync + 'static, 

Configure the actions for a fail point at runtime.

Each fail point can be configured by a callback. Process will call this callback function when it meet this fail-point.